Extract devices
